requirements.txt
autobahn
TikTokLive
shulker
python-dotenv